+++ docs/9_Inference_TensoRT_Models_on_ROS2_using_Cpp.md
@@ -1,11 +1,5 @@
-## Install python-wrapped centerpoint for tensorRT.
-``` shell
-cd ~/PerceptionRT/tools/tensorrt
-cmake -BRelease
-cmake --build Release
-```
 
-## (Python) Inference a model with TensorRT on ROS2 Node
+## (C++) Inference a model with TensorRT on ROS2 Node
 
 ### 1. Setting ROS2
 - If you are not using a container env, please install ROS2 Foxy:
@@ -23,26 +17,41 @@ sudo pip install transforms3d -y
 
 - If you are using a container env, your container already installed ROS2.
 
-### 2. Convert Waymo dataset to Rosbag for ROS2
-- You can use two repositories step by step. 
-  - Waymo data ==> Rosbag for ROS1: https://github.com/yukke42/waymo2bag
-    - 첫번째 깃허브로 ROS1 bag 파일로 만든 후
-  - Rosbag for ROS1 ==> Rosbag for ROS2: https://ternaris.gitlab.io/rosbags/topics/convert.html
-    - 2번째 라이브러리로 ros1을 ros2 bag파일로 변환하면 됨
-  
+### 2. Build the ROS2 package of centerpoint c++ Node in your ROS2 workspace.
+``` shell
+docker exec -it lidar3d-RT bash
+cd ~/ && mkdir -p ros2_ws/src && cd ros2_ws/ && colcon build --symlink-install
+cd src && ln -s ~/PerceptionRT/tools/ros2/centerpoint/ .
+```
+- If not exist `centerpoint/models`, copy `onnx` file.
+``` shell
+cd centerpoint && mkdir models
+cd models
+ln -s ~/PerceptionRT/onnx/config.yaml config.yaml
+ln -s ~/PerceptionRT/onnx/model.onnx model.onnx  
 
-### 3. ROS2 play bagfile on the container
+ln -s ~/PerceptionRT/onnx/model.onnx model.onnx
+```
+- Build the ROS2 package of centerpoint c++ Node in your ROS2 workspace.
+``` shell
+cd ~/ros2_ws && colcon build --symlink-install
+source ~/ros2_ws/install/setup.bash
+```
+
+### 3. Run the ROS2 Node.
+``` shell
+docker exec -it lidar3d-RT bash
+source ~/ros2_ws/install/setup.bash
+ros2 launch centerpoint centerpoint.launch.py
+```
+
+- Once running ros2 centerpoint node, create tensorRT file to the same folder having onnx file, automatically.
+
+### 4. ROS2 play bagfile on the container
 ```
 docker exec -it lidar3d-RT bash
 cd /Dataset
 ros2 bag play segment-10359308928573410754_720_000_740_000_with_camera_labels/  # ros2 bag play folder_with_ros2bag
-```
-
-### 4. Run ros2_demo.py on the container
-``` shell
-docker exec -it lidar3d-RT bash
-cd ~/PerceptionRT/tools/
-python ros2_demo.py --cfg_file cfgs/waymo_models/centerpoint_pillar_inference.yaml --ckpt ../ckpt/checkpoint_epoch_24.pth
 ```
 
 ### 5. Run rviz2
@@ -50,15 +59,10 @@ python ros2_demo.py --cfg_file cfgs/waymo_models/centerpoint_pillar_inference.ya
 docker exec -it lidar3d-RT bash
 rviz2
 ```
-
-### 6. Setting rviz2
 - Fixed Frame: base_link
 - Add -> By display type -> PountCloud2 -> Topic: /lidar/top/pointcloud, Size(m): 0.03
 - Add -> By topic -> /boxes/MarkerArray
 
-<img src="../sources/rviz2_add_topic.png" align="center" width="359">
-
-### 7. Results with rviz2
-<img src="./sources/rviz2.png" align="center" width="100%">
+<img src="./sources/rviz2_ros_cpp.png" align="center" width="100%">
 
 ## [Return to the main page.](../README.md)
