+++ general/config/config.py
@@ -88,4 +88,5 @@ def cfg_from_yaml_file(cfg_file, config):
 
 cfg = EasyDict()
 cfg.ROOT_DIR = (Path(__file__).resolve().parent / '../../').resolve()
+print(cfg.ROOT_DIR)
 cfg.LOCAL_RANK = 0
