--- tools/_init_path.py
@@ -1,2 +1,0 @@
-import sys
-sys.path.insert(0, '../')
